+++ src/init.c
@@ -145,7 +145,7 @@ extern r_obj* ffi_locate_interval_merge_groups(r_obj*, r_obj*, r_obj*, r_obj*);
 extern r_obj* ffi_interval_complement(r_obj*, r_obj*, r_obj*, r_obj*);
 extern r_obj* ffi_check_list(r_obj*, r_obj*);
 extern r_obj* ffi_is_list_of_vectors(r_obj*);
-extern r_obj* ffi_check_list_of_vectors(r_obj*, r_obj*);
+extern r_obj* ffi_list_check_all_vectors(r_obj*, r_obj*);
 
 
 // Maturing
@@ -312,7 +312,7 @@ static const R_CallMethodDef CallEntries[] = {
   {"ffi_interval_complement",               (DL_FUNC) &ffi_interval_complement, 4},
   {"ffi_check_list",                        (DL_FUNC) &ffi_check_list, 2},
   {"ffi_is_list_of_vectors",                (DL_FUNC) &ffi_is_list_of_vectors, 1},
-  {"ffi_check_list_of_vectors",             (DL_FUNC) &ffi_check_list_of_vectors, 2},
+  {"ffi_list_check_all_vectors",            (DL_FUNC) &ffi_list_check_all_vectors, 2},
   {NULL, NULL, 0}
 };