+++ NEWS.md
@@ -2,7 +2,7 @@
 
 * vctrs now depends on cli to help with error message generation.
 
-* New `vec_check_list()` and `vec_check_list_of_vectors()` input
+* New `vec_check_list()` and `list_check_all_vectors()` input
   checkers, and an accompanying `vec_is_list_of_vectors()` predicate.
 
 * New `vec_interleave()` for combining multiple vectors together, interleaving